pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ollama-vision-mcp"
version = "2.0.0"
description = "A Model Context Protocol server providing computer vision capabilities using Ollama"
authors = [
    {name = "Ollama Vision MCP Contributors"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
    "Topic :: Scientific/Engineering :: Image Recognition",
    "Topic :: Scientific/Engineering :: Artificial Intelligence"
]
keywords = ["mcp", "ollama", "vision", "computer-vision", "ai", "llm"]

dependencies = [
    "mcp>=0.9.1",
    "fastmcp>=0.2.0",
    "aiohttp>=3.8.0",
    "aiofiles>=23.0.0",
    "Pillow>=10.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.12.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0"
]

[project.urls]
"Homepage" = "https://github.com/ollama-vision-mcp/ollama-vision-mcp"
"Bug Reports" = "https://github.com/ollama-vision-mcp/ollama-vision-mcp/issues"
"Source" = "https://github.com/ollama-vision-mcp/ollama-vision-mcp"

[project.scripts]
ollama-vision-mcp = "src.server:main"

[tool.setuptools]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning"
]
